Make sphere unblocking + move to data folder
